Show territory map in scramble landing dialog.
This change re-uses the UI for retreat for scrambling landing, which brings the benefit of showing a map showing the territory being selected. Moves RetreatComponent to its own file, renaming it to SelectTerritoryComponent.
